Clean Up insides of GenerationSurfaceCollection
GenerationSurfaceCollection is now a dict of lists of NamedTuples.
GenerationSurface is replaced by a function generation_surface() which
creates a GenerationSurfaceCollection with a single surface.
This cleans up a lot of very anoying duck typing code which had to
treat a surface and a collection as the same thing. This should make
typing a lot easier