More versatile generation surfaces
Before this commit the generation surface was assumed to have 3
components: an event_weight a spatial pdf and an energy pdf. This mostly
worked but had some problems. There was weirdness caused by the fact
that genie doesn't store the zenith angle so it had to be faked.
This is also hindering future developement for more types of simulation.
This commit changes it so that generation surface is now an arbitrary
length sequence of pdfs, such that any amount of pdfs can be used with
any name columns.