Add ARG005 in ruff ignores
For the test of weighted effective area, I want to use lambda functions. The calculation of the integral of the flux demands a pdgid argument which however is actually not needed for a primary-independent flux weighting. So pdgid needs to be defined as argument of the flux function but not used which usually ruff would complain about.
I hope it's okay when I ignore ARG005