Adding another effective area weighting test (#53)
* Adding another effective area weighting test

Adding a test for the weighted version of effective area with a lambda function for an E^-2.7 spectrum. Note the pdgid necessary for get_weights()

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add ARG005 in ruff ignores

For the test of weighted effective area, I want to use lambda functions. The calculation of the integral of the flux demands a pdgid argument which however is actually not needed for a primary-independent flux weighting. So pdgid needs to be defined as argument of the flux function but not used which usually ruff would complain about.
I hope it's okay when I ignore ARG005

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>