Add standard deviation to the effective area function (#54)
* Add standard deviation to the effective area function

Added the calculation of the standard deviation of the effective area by weighting with squared weights. The default behavior of the function is unchanged, the return type is a single histograms with effective area binned in energy and cos(zenith). If the new argument return_stddev is set to True, a tuple of effective area and its uncertainty is returned.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Addressing mypy and ruff complaints

Making return_stddev a keyword-only argument to address FBT001 and FBT002 and changing from Tuple to tuple (UP006)

* Whoops

undo unintended change

* Trying to appease mypy with forward-declaring output type

let's see if it works now

* Add test for Aeff uncertainty

Adding a test for the standard deviation of effective area

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>