Make PDFs more explicit
One of the criticistms of simweights is that it is hard to find the
weighting functions. This was because the weighting was a list of
objects and the weights were calculating by taking the product of them.
This PR refactors it so that each type of simulation has a class that
is derived from GenerationSurface that overrides get_epdf(). This is a
clear function that is 1/weight. However since it makes use of Spatial
and PowerLaw it is not as explicit as it could be but DRY. As a side
effecth Spatial disturcitions and PowerLaw no longer have column names
as a member variable which i thought was just weird.